initial commit for tls_lab playbook and associated roles
The tls_lab playbook is going to be used for some blog posts to set up and
configure internode and client-to-node encryption.

The cassandra_ca role generates a CA for signing node certificates and it also
creates a truststore.

The node_keystores role generate signed certs and keystores. It copies the
keystores to the nodes.

The cassandra_configuration role configures cassandra.yaml and copies it to
each node.

The tls_common role has some shared code that is reused by the other roles.